#1: Covered all the items requested in the issue.
All the items except for the Dockerfile and containerization aspects of the project were completed.

It would seem that Docker provides a shallow control over the filesystem and wouldn't allow easily interacting with file download, installs and IO operations. Had a discussion with a teammate on the topic over at Divisions Maintenance Group and that was enough for me to decide to step away from Docker for the project until someone can show me without a shed of a doubt that it's not only helpful, but also better than the alternative approach of providing the set of binaries in a zip file and distribute that to developers.