#2: Defined the fundamental domain of a system tool.
Without getting in the specifics of how we handle a system tool, let's rather focus on how we want to represent a system tool. A system tool is going to be a development utility (ex: Git) that's going to be commonly used by a developer. Such an entity would need multiple characteristics to describe what it is such as:
- The name of the tool (title),
- Its current semantic version,
- A high-level description if necessary,
- The selected mechanism to download the tool from the web
- Its absolute path on the machine,
- Whether it should be treated as an environment variable.

As we move further in the implementation, we'll be using the domain knowledge acquired here in the infrastructure and define the necessary code to download system tools from the web. Moreover, we'll also have to define an easy-going framework for the users to use if they want to add new default system tools to DBS.